+++ apps/app/src/app.controller.ts
@@ -7,8 +7,14 @@ import {
 } from "@nestjs/common";
 import { AppService } from "./app.service";
 import { SignType } from "./types/sign-types.enum";
-import { GenerateRequestDto, KeyRequestDto, RawSignRequestDto, SignRequestDto } from "./types/request.dto";
+import {
+  GenerateRequestDto,
+  KeyRequestDto,
+  RawSignRequestDto,
+  SignRequestDto,
+} from "./types/request.dto";
 import { EncryptedPayloadDto } from "./types/encrypted-payload.dto";
+import { RequestBodyValidationPipe } from "./pipes/request-body-validation.pipe";
 
 @Controller()
 export class AppController {
@@ -17,7 +23,8 @@ export class AppController {
   @Post("sign/vc/:type")
   async signCredential(
     @Param("type", new ParseEnumPipe(SignType)) type: SignType,
-    @Body() body: SignRequestDto | EncryptedPayloadDto
+    @Body(new RequestBodyValidationPipe(SignRequestDto))
+    body: SignRequestDto | EncryptedPayloadDto
   ) {
     return await this.appService.signCredential(type, body);
   }
@@ -25,7 +32,8 @@ export class AppController {
   @Post("sign/vp/:type")
   async signPresentation(
     @Param("type", new ParseEnumPipe(SignType)) type: SignType,
-    @Body() body: SignRequestDto | EncryptedPayloadDto
+    @Body(new RequestBodyValidationPipe(SignRequestDto))
+    body: SignRequestDto | EncryptedPayloadDto
   ) {
     return await this.appService.signPresentation(type, body);
   }
@@ -33,23 +41,33 @@ export class AppController {
   @Post("sign/pop/:type")
   async signProofOfPossession(
     @Param("type", new ParseEnumPipe(SignType)) type: SignType,
-    @Body() body: SignRequestDto | EncryptedPayloadDto,
+    @Body(new RequestBodyValidationPipe(SignRequestDto))
+    body: SignRequestDto | EncryptedPayloadDto
   ) {
     return await this.appService.signProofOfPossession(type, body);
   }
 
   @Post("sign/raw")
-  async signRaw(@Body() body: RawSignRequestDto | EncryptedPayloadDto) {
+  async signRaw(
+    @Body(new RequestBodyValidationPipe(RawSignRequestDto))
+    body: RawSignRequestDto | EncryptedPayloadDto
+  ) {
     return await this.appService.signRaw(body);
   }
 
   @Post("generate")
-  generateKey(@Body() body: GenerateRequestDto | EncryptedPayloadDto) {
+  generateKey(
+    @Body(new RequestBodyValidationPipe(GenerateRequestDto))
+    body: GenerateRequestDto | EncryptedPayloadDto
+  ) {
     return this.appService.generateKey(body);
   }
 
   @Post("delete")
-  deleteKey(@Body() body: KeyRequestDto | EncryptedPayloadDto) {
+  deleteKey(
+    @Body(new RequestBodyValidationPipe(KeyRequestDto))
+    body: KeyRequestDto | EncryptedPayloadDto
+  ) {
     return this.appService.deleteKey(body);
   }
 }