+++ apps/app/test/app.e2e-spec.ts
@@ -15,10 +15,11 @@ describe("AppController (e2e)", () => {
     await app.init();
   });
 
-  it("/ (GET)", () => {
-    return request(app.getHttpServer())
-      .get("/")
-      .expect(200)
-      .expect("Hello World!");
+  afterEach(async () => {
+    await app.close();
+  });
+
+  it("GET /health", () => {
+    return request(app.getHttpServer()).get("/health").expect(200);
   });
 });