+++ apps/app/test/test-setup.ts
@@ -1,4 +1,5 @@
 // Global test setup
+import "reflect-metadata";
 import { jest } from "@jest/globals";
 import { webcrypto } from "node:crypto";
 import * as crypto from "node:crypto";